MAILBOX-351 Factorize ReIndexing tasks details
There was a lot of unnecessary code duplication
