+++ mailbox/tools/indexer/src/main/java/org/apache/mailbox/tools/indexer/SingleMailboxReindexingTask.java
@@ -23,52 +23,24 @@ import java.util.Optional;
 
 import javax.inject.Inject;
 
-import org.apache.james.mailbox.indexer.IndexingDetailInformation;
-import org.apache.james.mailbox.indexer.ReIndexingExecutionFailures;
 import org.apache.james.mailbox.model.MailboxId;
 import org.apache.james.task.Task;
 import org.apache.james.task.TaskExecutionDetails;
-
-import com.fasterxml.jackson.annotation.JsonIgnore;
-import com.fasterxml.jackson.annotation.JsonProperty;
 
 public class SingleMailboxReindexingTask implements Task {
 
     public static final String MAILBOX_RE_INDEXING = "mailboxReIndexing";
 
-    public static class AdditionalInformation implements TaskExecutionDetails.AdditionalInformation, IndexingDetailInformation {
+    public static class AdditionalInformation extends ReprocessingContextInformation {
         private final MailboxId mailboxId;
-        private final ReprocessingContext reprocessingContext;
 
         AdditionalInformation(MailboxId mailboxId, ReprocessingContext reprocessingContext) {
+            super(reprocessingContext);
             this.mailboxId = mailboxId;
-            this.reprocessingContext = reprocessingContext;
         }
-
 
         public String getMailboxId() {
             return mailboxId.serialize();
-        }
-
-        @Override
-        public int getSuccessfullyReprocessMailCount() {
-            return reprocessingContext.successfullyReprocessedMailCount();
-        }
-
-        @Override
-        public int getFailedReprocessedMailCount() {
-            return reprocessingContext.failedReprocessingMailCount();
-        }
-
-        @Override
-        @JsonIgnore
-        public ReIndexingExecutionFailures failures() {
-            return reprocessingContext.failures();
-        }
-
-        @JsonProperty("failures")
-        public SerializableReIndexingExecutionFailures failuresAsJson() {
-            return SerializableReIndexingExecutionFailures.from(failures());
         }
     }
